% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig.R
\name{ig}
\alias{ig}
\title{Entropy and Information Gain}
\usage{
ig(root, df, ignore)
}
\arguments{
\item{root}{The parent root variable.}

\item{df}{Your dataset or dataframe.}

\item{ignore}{The variables or columns in your dataset that you don't want to include}
}
\value{
Output is a dataframe with all the variables' entropy and information gain.
}
\description{
\code{library(shannon.entrpoy)}
\code{ig()} returns the entropy and information gains of all the variables present in your dataset.
}
\details{
Important note: if you have continuous variables, be sure to handle that separately or else you will get useless results.
Recommended to split or cut your continuous variables into bins
}
\examples{
ig(root='purchase', df=myData, ignore=c('id','name'))
ig(root='credit', df=salesdata, ignore='age')

}
